[MSA-99] refactor : Company Service수정중
