+++ company/src/main/java/com/i4u/company/domain/repository/CompanyRepository.java
@@ -3,17 +3,12 @@ package com.i4u.company.domain.repository;
 import com.i4u.company.domain.Company;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 import java.util.UUID;
 
-public interface CompanyRepository {
-
-    Company save(Company company);
-
-    Optional<Company> findById(UUID companyId);
-
-    Page<Company> findByName(String keyword, Pageable pageable);
-
-    void softDeleteCompany(UUID companyId);
+@Repository
+public interface CompanyRepository extends JpaRepository<Company, UUID> {
 }
