+++ company/src/main/java/com/i4u/company/infrastructure/CompanyRepositoryImpl.java
@@ -15,53 +15,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@Repository
-@RequiredArgsConstructor
-public class CompanyRepositoryImpl implements CompanyRepository {
+public abstract class CompanyRepositoryImpl implements CompanyRepository {
 
-    private final CompanyJpaRepository jpaRepository;
-    private final JPAQueryFactory queryFactory;
-
-    @Override
-    public Company save(Company company) {
-        return jpaRepository.save(company);
-    }
-
-    @Override
-    public Optional<Company> findById(UUID id) {
-        return jpaRepository.findById(id);
-    }
-
-    @Override
-    public Page<Company> findByName(String keyword, Pageable pageable) {
-        List<Company> companies = queryFactory
-                .selectFrom(company)
-                .where(
-                        company.name.value.containsIgnoreCase(keyword), // 대소문자 무시
-                        company.isDeleted.eq(false) // 삭제되지 않은 가게만 조회
-                )
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize())
-                .fetch();
-
-        Long total = queryFactory
-                .select(company.count())
-                .from(company)
-                .where(
-                        company.name.value.containsIgnoreCase(keyword),
-                        company.isDeleted.eq(false) // 삭제되지 않은 가게만 조회
-                )
-                .fetchOne();
-        return new PageImpl<>(companies, pageable, total != null ? total : 0);
-    }
-
-    @Override
-    public void softDeleteCompany(UUID companyId) {
-        queryFactory
-                .update(company)
-                .set(company.isDeleted, true)
-                .where(company.id.eq(companyId))
-                .execute();
-
-    }
 }
